requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
astral==3.2
attrs==24.2.0
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorzero==2.0
cryptography==44.0.0
dnspython==2.7.0
eventlet==0.38.0
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
gpiozero==2.0.1
greenlet==3.1.1
h11==0.14.0
http_ece==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
opencv-python==4.10.0.84
propcache==0.2.1
psutil==6.1.0
py-vapid==1.9.2
pycparser==2.22
pyftdi==0.56.0
pyserial==3.5
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
pyusb==1.2.1
pywebpush==2.0.3
requests==2.32.3
ruamel.yaml==0.18.6
setuptools==75.6.0
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
